+++ lib/edge_ai.dart
@@ -10,4 +10,8 @@ export 'edge_gen_ai_rewrite_style.dart' show EdgeGenAIRewriteStyle;
 export 'edge_gen_ai_rewriter.dart' show EdgeGenAIRewriter;
 export 'edge_gen_ai_summarizer.dart' show EdgeGenAISummarizer;
 export 'edge_gen_ai_tool.dart'
-    show EdgeGenAITool, EdgeGenAIToolParameter, EdgeGenAIToolParameterType;
+    show
+        EdgeGenAITool,
+        EdgeGenAIToolParameter,
+        EdgeGenAIToolParameterType,
+        EdgeGenAIToolSchema;